refactor(action): move action definition from character to action module
