+++ src/.gitignore
@@ -1,4 +1,5 @@
 !action
+!character
 !data
 !roll
